requirements.txt
polars==0.20.21
youtube_transcript_api==0.6.2
sentence-transformers==2.6.1
requests==2.31.0
google-api-python-client==2.156.0
PyYAML==6.0.2
python-dotenv==1.0.1
numpy==2.0.2
pandas==2.2.3
# -e .